+++ src/games/index.ts
@@ -12,5 +12,6 @@ import { expeditions } from './expeditions'
 import { frosthaven } from './frosthaven'
 import { nemesis } from './nemesis'
 import { taintedgrail } from './taintedgrail'
+import { taintedgrailkor } from './taintedgrailkor'
 
-export const TUTORIALS: Tutorial[] = [nemesis, taintedgrail, expeditions, frosthaven, bitoku]
+export const TUTORIALS: Tutorial[] = [nemesis, taintedgrail, taintedgrailkor, expeditions, frosthaven, bitoku]